Refactor: KakaoProperties =>properties [박한솔]
