+++ web3-credential-server/src/main/java/web3/Application.java
@@ -5,7 +5,7 @@ import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.boot.context.properties.ConfigurationPropertiesScan;
 
 @SpringBootApplication
-@ConfigurationPropertiesScan("web3.controller.kakao")
+@ConfigurationPropertiesScan("web3.properties.kakao")
 public class Application {
     public static void main(String[] args) {
         SpringApplication.run(Application.class, args);
