+++ web3-credential-server/src/main/java/web3/controller/kakao/AdminKakaoController.java
@@ -4,6 +4,7 @@ import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
+import web3.properties.kakao.KakaoProperties;
 
 @Controller
 @RequestMapping("/admin/kakao")
