+++ web3-credential-server/src/main/java/web3/properties/kakao/KakaoProperties.java
@@ -1,4 +1,4 @@
-package web3.controller.kakao;
+package web3.properties.kakao;
 
 import org.springframework.boot.context.properties.ConfigurationProperties;
