+++ web3-credential-server/src/main/java/web3/service/kakao/KakaoService.java
@@ -12,9 +12,9 @@ import org.springframework.util.LinkedMultiValueMap;
 import org.springframework.web.client.RestClient;
 import org.springframework.web.client.RestClientException;
 import org.springframework.web.client.RestClientResponseException;
-import web3.controller.kakao.KakaoProperties;
+import web3.properties.kakao.KakaoProperties;
 import web3.domain.user.User;
-import web3.domain.user.dto.UserInfoDto;
+import web3.common.dto.user.UserInfoDto;
 import web3.domain.wallet.Wallet;
 import web3.repository.wallet.WalletRepository;
 import web3.service.user.UserService;
